requirements.txt
streamlit==1.47.0
pymupdf==1.26.3
nltk==3.8.1
spacy==3.8.7
sentence-transformers==5.0.0
matplotlib==3.9.0
seaborn==0.13.2
faiss-cpu==1.11.0
numpy==1.26.4
pandas==2.1.4
regex==2024.11.6


